// src/api/models/Register-model.ts
export interface RegisterClientModel {
    email: string
    password: string
    firstname: string
    lastname: string
    phone: string
    city: string
}

export interface RegisterProModel {
    email: string
    password: string
    firstname: string
    lastname: string
    phone: string
    city: string
    diplome_pro: string
    price_pro: string
    city_pro: string,
    country_pro: string,
    phone_pro: string,
    email_pro:string,
    address_pro: string,
    siren_pro: string,
    siret_pro:string,
    title_pro: string,
    description_pro: string,
}